Refactor nearpc
